requirements-dev.txt
black[jupyter]
build
isort
pre-commit
pycln
pytest-cov
twine
autodoc_pydantic
parameterized